Extracted view state and events to separate files
