+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/payments/SelectPaymentMethodViewState.kt
@@ -1,0 +1,19 @@
+package com.woocommerce.android.ui.payments
+
+import com.woocommerce.android.model.UiString
+
+sealed class SelectPaymentMethodViewState {
+    object Loading : SelectPaymentMethodViewState()
+    data class Success(
+        val paymentUrl: String,
+        val orderTotal: String,
+        val isPaymentCollectableWithExternalCardReader: Boolean,
+        val isPaymentCollectableWithTapToPay: Boolean,
+        val learMoreIpp: LearMoreIpp,
+    ) : SelectPaymentMethodViewState()
+
+    data class LearMoreIpp(
+        val label: UiString,
+        val onClick: () -> Unit,
+    )
+}
